requirements.txt
google-cloud-documentai
google-auth
